[refactor] #19 : mapData 타입 분리
