+++ src/shared/stores/useMapStore.ts
@@ -1,55 +1,11 @@
-import { create } from 'zustand';
-import { mockMapData } from '@/shared/model';
+import { create } from "zustand";
+import { mockMapData } from "@/shared/model";
+import { MapState } from "../model/MapData.type";
 
-interface TransferInfo {
-  type: string;
-  duration: string;
-  line?: string;
-  board?: string;
-  stations?: string[];
-  distance?: number;
-}
-
-interface DriveInfo {
-  driveDuration: string;
-  driveDistance: number;
-  toll: number;
-  taxiFareEstimate: number;
-}
-
-interface User {
-  id: string;
-  name: string;
-  profileImg?: string;
-  startStation: string;
-  latitude: number;
-  longitude: number;
-  transferInfo: TransferInfo[];
-  exit: string;
-  destination: string;
-  totalTime: string;
-  drive?: DriveInfo;
-}
-
-interface MeetingPoint {
-  stationName: string;
-  latitude: number;
-  longitude: number;
-  averageDuration: string;
-}
-
-interface MapState {
-  users: User[];
-  meetingPoint: MeetingPoint;
-
-  setUsers: (users: User[]) => void;
-  setMeetingPoint: (meetingPoint: MeetingPoint) => void;
-}
-
-export const useMapStore = create<MapState>((set) => ({
+export const useMapStore = create<MapState>(set => ({
   users: mockMapData.users,
   meetingPoint: mockMapData.meetingPoint,
 
-  setUsers: (users) => set({ users }),
-  setMeetingPoint: (meetingPoint) => set({ meetingPoint }),
+  setUsers: users => set({ users }),
+  setMeetingPoint: meetingPoint => set({ meetingPoint }),
 }));
